+++ src/pages/clips.tsx
@@ -5,6 +5,7 @@ import { useRouter } from 'next/router';
 import React, { useEffect, useState } from 'react';
 import toast from 'react-hot-toast';
 import { SyncLoader } from 'react-spinners';
+import { ClipWithPreview } from 'src/typings/interclip';
 
 interface ClipsResponse {
   status: 'error' | 'success';
