+++ src/pages/file.tsx
@@ -22,7 +22,7 @@ export default function HomePage() {
     try {
       const fileURL = await uploadFile(filesEndpoint, e);
       const clipResponse = await requestClip(fileURL);
-      if (clipResponse) {
+      if (clipResponse && clipResponse.status === 'success') {
         const { result: clip } = clipResponse;
         setCode(clip.code.slice(0, clip.hashLength));
         setUploaded(true);
