+++ src/typings/wikipedia.d.ts
@@ -1,0 +1,23 @@
+interface WikipediaArticle {
+  pageid: number;
+  ns: number;
+  title: string;
+  images: {
+    ns: number;
+    description: string;
+  }[];
+}
+
+interface WikipediaResponse {
+  batchcomplete: string;
+  continue: {
+    grncontinue: string;
+    continue: string;
+  };
+  warnings: any;
+  query: {
+    pages: {
+      [key: string]: WikipediaArticle;
+    };
+  };
+}
